Eliminated race condition to do with assigning WebSocket event handlers
In JavaScript, a websocket's event handlers are assigned to it after creation/connection.
Normally, this is not a problem as JavaScript's single threaded nature requires code in the current
context to finish before it will start accepting the events.

In my implementation, I was calling out to the AHK function _SaveWS before assigning
the event handlers. This evidently caused some kind of context switch allowing JavaScript
to start accepting events before the handlers were all in place.

By accessing the websocket through the DOM rather than the previous method the need to
call out to _SaveWS is eliminated. By that, so is the race condition.